Adapted to language changes.

+++ src/data_structures/fmindex.rs
@@ -229,6 +229,7 @@ impl<'a> FMDIndex<'a> {
             curr.clear();
             // size of the last confirmed interval
             let mut last_size = -1;
+            
             // iterate over forward extensions in reverse, as they are sorted by size
             // and we prefer longer matches
             for &interval in prev.iter().rev() {
@@ -245,8 +246,8 @@ impl<'a> FMDIndex<'a> {
                     }
                 }
                 // add _interval to curr (will be further extended next iteration)
-                if _interval.size != 0 && _interval.size != last_size {
-                    last_size = _interval.size;
+                if _interval.size != 0 && _interval.size as isize != last_size {
+                    last_size = _interval.size as isize;
                     curr.push(_interval);
                 }
             }
